requirements.txt
django~=4.2.2
djangorestframework~=3.14.0
django-filter~=23.2
psycopg2-binary

pytest~=7.3.2